+++ src/client/components/index.ts
@@ -1,4 +1,0 @@
-export * from './hljs';
-export * from './icon';
-export * from './rfc7230';
-export * from './testResults';