+++ src/client/css.d.ts
@@ -1,4 +1,0 @@
-declare module '*.css' {
-  const classes: Record<string, string>;
-  export default classes;
-}