+++ src/extension/httpOutputProvider/index.ts
@@ -1,6 +1,6 @@
-export * from './builtInHttpOutputProvider';
 export * from './contentTypeHttpOutputProvider';
 export * from './imageHttpOutputProvider';
 export * from './markdownHttpOutputProvider';
+export * from './monacoEditorHttpOutputProvider';
 export * from './rfc7230HttpOutputProvider';
 export * from './testResultsHttpOutputProvider';