+++ src/renderer/components/testResults.tsx
@@ -1,7 +1,7 @@
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import { h, FunctionComponent, Fragment } from 'preact';
-import style from './testResults.css';
-import { TestResult } from 'httpyac';
+import './testResults.css';
+import type { TestResult } from 'httpyac';
 import PassSvg from 'vscode-codicons/src/icons/pass.svg';
 import BugSvg from 'vscode-codicons/src/icons/bug.svg';
 import { Icon } from '../components/icon';
@@ -9,13 +9,13 @@ import { Icon } from '../components/icon';
 export const TestResults: FunctionComponent<{
   testResults: Array<TestResult>;
 }> = ({ testResults }) => (
-  <div className={style.testResults}>
-    <div className={style.testTitle}>
-      <h4 className={style.testTitleCell}>
+  <div class="testresults">
+    <div class="testtitle">
+      <h4 class="testtitle__cell">
         <span>{testResults.filter(obj => obj.result).length}</span> Passing Tests
       </h4>
-      <h4 className={style.testTitleCell}>
-        <span className={ testResults.filter(obj => !obj.result).length > 0 ? style.testError : ''}>{testResults.filter(obj => !obj.result).length}</span> Failing Tests
+      <h4 class="testtitle__cell">
+        <span className={ testResults.filter(obj => !obj.result).length > 0 ? 'is_error' : ''}>{testResults.filter(obj => !obj.result).length}</span> Failing Tests
       </h4>
     </div>
     <section>
@@ -29,13 +29,11 @@ export const TestResults: FunctionComponent<{
 
 const TestResultLine: FunctionComponent<{ testResult: TestResult }> = ({ testResult }) => (
 
-  <article className={style.testLine}>
-    <span className={testResult.result ? '' : style.testError}>
-      {testResult.result
-        ? <Icon svg={PassSvg}></Icon>
-        : <Icon svg={BugSvg}></Icon>
-      }
-    </span>
+  <article class="testline">
+    {testResult.result
+      ? <Icon svg={PassSvg} class="is_success"></Icon>
+      : <Icon svg={BugSvg} class="is_error"></Icon>
+    }
     <div>
       <span>{testResult.message}</span>
       {
@@ -43,7 +41,7 @@ const TestResultLine: FunctionComponent<{ testResult: TestResult }> = ({ testRes
         && <Fragment>
           <div>{testResult.error.displayMessage}</div>
           {
-            testResult.error?.error?.stack && <pre><code className={style.testError}>{testResult.error.error.stack}</code></pre>
+            testResult.error?.error?.stack && <pre><code class="testline__error">{testResult.error.error.stack}</code></pre>
           }
         </Fragment>
       }