+++ webpack.config.js
@@ -1,3 +1,4 @@
+const MonacoWebpackPlugin = require('monaco-editor-webpack-plugin');
 const ForkTsCheckerWebpackPlugin = require('fork-ts-checker-webpack-plugin');
 const { DefinePlugin } = require('webpack');
 const path = require('path');
@@ -21,7 +22,8 @@ module.exports = [(env, argv) => {
     },
     devtool: argv.mode === 'development' ? 'eval-cheap-module-source-map' : 'nosources-source-map',
     externals: {
-      vscode: 'commonjs vscode'
+      vscode: 'commonjs vscode',
+      httpyac: 'httpyac'
     },
     resolve: {
       extensions: ['.ts', '.js']
@@ -64,85 +66,110 @@ module.exports = [(env, argv) => {
     },
   };
   return config;
-}, (env, argv) => ({
-  mode: argv.mode,
-  devtool: argv.mode === 'production' ? false : 'inline-source-map',
-  entry: './src/client/index.ts',
-  output: {
-    path: path.join(__dirname, 'dist'),
-    filename: "client.js",
-    publicPath: '',
-  },
-  resolve: {
-    extensions: ['.ts', '.tsx', '.js', '.jsx', '.css'],
-  },
-  module: {
-    rules: [
-      {
-        test: /\.tsx?$/,
-        use: [
-          {
-            loader: 'thread-loader',
-          },
-          {
-            loader: 'ts-loader',
-            options: {
-              happyPackMode: true,
-              configFile: 'src/client/tsconfig.json',
-              transpileOnly: true,
-              compilerOptions: {
-                noEmit: false,
+}, (env, argv) => {
+  /**@type {import('webpack').Configuration}*/
+  const config = {
+    mode: argv.mode,
+    devtool: argv.mode === 'development' ? 'eval-cheap-module-source-map' : false,
+    entry: {
+      monacoRenderer: './src/renderer/monacoRenderer.tsx',
+      testResultsRenderer: './src/renderer/testResultsRenderer.tsx',
+      rfc7230Renderer: './src/renderer/rfc7230Renderer.tsx',
+    },
+    output: {
+      path: path.join(__dirname, 'dist'),
+      filename: "[name].js",
+      libraryTarget: 'module',
+    },
+    experiments: {
+      outputModule: true,
+    },
+    resolve: {
+      extensions: ['.ts', '.tsx', '.js', '.jsx', '.css'],
+    },
+    module: {
+      rules: [
+        {
+          test: /\.tsx?$/,
+          use: [
+            {
+              loader: 'thread-loader',
+            },
+            {
+              loader: 'ts-loader',
+              options: {
+                happyPackMode: true,
+                configFile: 'src/renderer/tsconfig.json',
+                transpileOnly: true,
+                compilerOptions: {
+                  noEmit: false,
+                },
               },
-            },
-          }
-        ]
+            }
+          ]
 
-      },
-      {
-        test: /\.css$/,
-        use: [
-          'style-loader',
-          {
-            loader: 'css-loader',
-            options: {
-              importLoaders: 1,
-              modules: true,
-            },
-          },
-        ],
-      },
-			{
-				test: /\.svg$/,
-				loader: 'svg-sprite-loader',
-			},
-    ],
-  },
-  devServer: {
-    port: devServerPort,
-    hot: true,
-    disableHostCheck: true,
-    writeToDisk: true,
-    headers: { 'Access-Control-Allow-Origin': '*' },
-  },
-  plugins: [
-    new ForkTsCheckerWebpackPlugin({
-      async: true,
-      typescript: {
-        tsconfig: 'src/client/tsconfig.json',
-        diagnosticOptions: {
-          semantic: true,
-          syntactic: true,
         },
-      },
-      eslint: {
-        files: ['./src/client/**/*.{ts,tsx,js,jsx}']
-      }
-    }),
-    new DefinePlugin({
-      // Path from the output filename to the output directory
-      __webpack_relative_entrypoint_to_root__: JSON.stringify(
-        path.posix.relative(path.posix.dirname(`/client.js`), '/'),
-      ),
-    }),
-  ],
-})];
+        {
+          test: /\.css$/,
+          use: [
+            'style-loader',
+            {
+              loader: 'css-loader'
+            },
+          ],
+        },
+        {
+          test: /\.svg$/,
+          loader: 'svg-sprite-loader',
+        },
+        {
+          test: /\.ttf$/,
+          use: ['file-loader']
+        }
+      ],
+    },
+    devServer: {
+      port: devServerPort,
+      hot: true,
+      disableHostCheck: true,
+      writeToDisk: true,
+      headers: { 'Access-Control-Allow-Origin': '*' },
+    },
+    plugins: [
+      new MonacoWebpackPlugin({
+        languages: ['css', 'graphql', 'html', 'ini', 'javascript', 'json', 'markdown', 'sql', 'xml', 'yaml'],
+        features: [
+          '!accessibilityHelp', '!anchorSelect', 'bracketMatching', '!caretOperations',
+          'clipboard', 'codeAction', '!codelens', '!colorPicker',
+          '!comment', 'contextmenu', '!coreCommands', '!cursorUndo',
+          '!dnd', '!documentSymbols', 'find', 'folding',
+          '!fontZoom', 'format', '!gotoError', '!gotoLine',
+          '!gotoSymbol', '!hover', '!iPadShowKeyboard', '!inPlaceReplace',
+          'indentation', '!inlineHints', '!inspectTokens', '!linesOperations',
+          '!linkedEditing', 'links', '!multicursor', '!parameterHints',
+          '!quickCommand', '!quickHelp', '!quickOutline', '!referenceSearch',
+          '!rename', '!smartSelect', '!snippets', '!suggest',
+          '!toggleHighContrast', '!toggleTabFocusMode', '!transpose', '!unusualLineTerminators',
+          '!viewportSemanticTokens', 'wordHighlighter', 'wordOperations', 'wordPartOperations'
+        ]
+      }),
+      new ForkTsCheckerWebpackPlugin({
+        async: true,
+        typescript: {
+          tsconfig: 'src/renderer/tsconfig.json',
+          diagnosticOptions: {
+            semantic: true,
+            syntactic: true,
+          },
+        },
+        eslint: {
+          files: ['./src/renderer/**/*.{ts,tsx,js,jsx}']
+        }
+      }),
+    ],
+    optimization: {
+      minimize: true
+    }
+  };
+  return config;
+}];